backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg==3.2.0
setuptools==70.3.0
sqlparse==0.5.0
typing_extensions==4.12.2
wheel==0.43.0
